requirements.txt
fastapi
uvicorn[standard]
transformers
torch
python-dotenv
accelerate
llama-cpp-python
aiosqlite
databases